requirements.txt
torch
torchvision
scikit-learn
numpy
pandas
matplotlib
streamlit
Pillow
scipy
seaborn
